2.19 Animaciones y detalles css

+++ src/components/notes/NoteScreen.jsx
@@ -1,5 +1,6 @@
 import React, { useEffect, useRef } from "react";
 import { useDispatch, useSelector } from "react-redux";
+import Swal from "sweetalert2";
 import { useForm } from "../../hooks/useForm";
 import { activeNote, startDelete } from "../actions/notes";
 import NoteAppBar from "./NoteAppBar";
@@ -27,11 +28,33 @@ const NoteScreen = () => {
 
   
   const handleDelete = ()=>{
-    dispatch(startDelete (id));
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        
+        dispatch(startDelete (id));
+        Swal.fire(
+          'Deleted!',
+          'Your file has been deleted.',
+          'success'
+        )
+      }
+    })
+    
+    
+    
+    
   }
   
   return (
-    <div className="notes__main-content">
+    <div className="notes__main-content animate__animated animate__fadeIn">
       <NoteAppBar />
 
       <div className="notes__content">
@@ -64,7 +87,7 @@ const NoteScreen = () => {
           <button 
             className="btn btn-danger"
             onClick={handleDelete}
-          >Delete
+          ><i class="far fa-trash-alt"></i> Borrar nota
           </button>
     
     </div>
